beaconing: fix bug in selection algorithm
In a refactor of the beacon selection aglorithm, a bug was introduced.

In case the most diverse beacon was already part of the resultSize-1
shortest beacons, a shallow copy of the most diverse beacon was appended
to the result slice, instead of the shortest beacon of the remaining beacons.

As a result, extending the beacon during segment termination for this
shallow copy fails, because we attempt to extend it twice. This could
even result in panics as observed in SCIONLab by @matzf.

Offending commit https://github.com/scionproto/scion/commit/aba5b81931b312714bc2567ac2e4a015173d5164

GitOrigin-RevId: 23eab382868da7b7c5fe4ba8b08e8c41b8aae711