dispatcher/snet: reduce packet buffer size
Excessively large packet buffer sizes were allocated in the dispatcher
and snet; for each connection, for each read and write a 256KB buffer
was allocated in the dispatcher, and 64KB in snet.
With e.g. the control service creating a large number of different
connections, the very large buffer allocations create a signficant
memory pressure, putting both the control service and the dispatcher
into RSS of around 1G.
The theoretical maximum MTU for SCION is 64KB, but only up to 9KB
ethernet jumbo frames make practical sense. In fact, the router only
supports 9KB packets.

Reduced the buffer size to 9000 - 20 - 8 (ethernet jumbo frame payload,
minus IPv4/UDP header size) in both in the dispatcher and snet.