scion-pki: add --check-time to certificate validate
Add the optional `--check-time` flag to `scion-pki certificate validate`
that checks whether the validated certificate covers the current time.
By default, we do not check the current time as validate is used to
check the profile of a certificate.

A certificate should only be trusted if it can be verified successfully,
which is done via the `scion-pki certificate verify` command.

GitOrigin-RevId: 81695d9043db9b9e0aed1872abc7668401b883e1