cs: remove noisy debug log message on beacon loop
Every single beacon loop detected with FilterLoop was logged from the
propagator. In setups with large number of candidate beacons, this is
extremely noisy, making the debug log quite unusable.
The loop filter appears to be simple and clear enough, so that having
this information in the log would rarely ever seem helpful.