cs: persist TRCs to certs subdirectory
The trustpersister configuration was such that TRCs were persisted to
the configuration directory root. Instead, they should be written
to the `certs/` subdirectory, where the loader will be looking for them.

According to @oncilla, the directory structure will eventually be changed
such that TRCs live in the `crypto/trcs/` subdirectory.

Closes #4100

GitOrigin-RevId: cd9ad1bdf364f021bdc74437de3133dc1babac9c