router: non-blocking write for packet forwarding
The structure of the router's dataplane is currently very simple, with a
single processing goroutine per (internal and external) router
interface. This goroutine reads and processes packets arriving on this
interface and finally writes it out to a (different) interface. This
write was a blocking call, i.e. the call would block until the packet
could be copied into the socket's sendbuffer. While the processing
goroutine is blocked, it cannot read and process packets, and incoming
packets will eventually be dropped.
With this setup, a slow (for writing) router interface can block
processing packets for all router interfaces, affecting traffic that is
not even destined to the slow interface.
Worse yet, with the polling based I/O implementation in golang, the
write may be blocked for longer than what would immediately be needed to
write the packet; in linux, poll only considers a socket as writable
once the sendbuffer has been drained to half. This can lead to long
stalls during which no packets can be processed.

Using a non-blocking write call when forwarding packets fixes these
issues and allows to keep the simple structure of the router's
dataplane.

 Additionally, remove the redundant underlay/conn/mock_conn.MockConn and
 use the router/mock_router.MockBatchConn in the router tests instead.

Closes #4096

GitOrigin-RevId: e5196e10240c7a08249cd6019bd4a8f0fd36eda5