deps: update rules_docker
Additionally, explicitly depend on https://github.com/google/go-containerregistry , which is a dependency of rules_docker.
rules_docker uses a tar file for the dependency. By the explicit dependency, we can leverage the regular go module proxy mechanism, which allows for better caching in CI.

Furthermore, remove the goproxy.io fallback proxy, as it is not need anymore.

GitOrigin-RevId: 41c2cab0f27120d3d9fdebd062ed9d82cb4808c4