app/env: add host-wide SCION environment file
This PR adds a SCION environment file that SCION applications can use to lookup the dispatcher socket, the daemon per ISD-AS and which source AS to use by default.

The environment can be specified in JSON. Currently, the contents are intentionally kept to a minimum, however, the json-encoding makes it very easy to expand in the future.

GitOrigin-RevId: ea1ad603f54fa57195b1dd507b1706bcc3b311e9