daemon: add periodic TRC loader
So far the daemon only loaded the TRC from the disk once and then it
would never have picked up a new TRC if it was added to the disk.

By adding a periodic task that loads the TRCs from disk we no longer need
to restart the daemon after adding new TRCs.

GitOrigin-RevId: f412b82215da108ccbd9054f45423d041fc427f5