cs: support running without active signer
Be less strict about requiring an active signer in the control service.
Because we support active reloading of the private key and certificate chains
it is reasonable to expect the service to run without them initially.

Most operations that are done by the control service do not require an active
signer, e.g., handling path requests. Only in AS entry creation it is required.

GitOrigin-RevId: b34f4c228105bc216ffa0522d8be321eac9a1484