cs: periodically load trcs from multiple directories
Add multiple optional directories where the TRCs can be loaded from
in the signer generator. We currently use the `crypto/as` and the
`certs` directories. `certs` is the place where currently TRCs are
expected to be located. `crypto/as` remains there to be backwards
compatible.

In the future, we can change `certs` to something more reasonable, and
phase out the `certs` and `crypto/as` directories without breaking change.

GitOrigin-RevId: fa0e9aaa44d04904f62dea4c37f7af3dcd0339a5