cs: properly load TRC in crypto loader wrapper
Load the TRC as part of the `SignedTRC` method instead of the `Chains` method
in the crypto loader wrapper for the trust database in the control service.

In the signer generator, we load TRCs before loading the chains. In case a control
service starts without a TRC available in the beginning, it would fail loading
the TRCs before it ever reaches the call to `Chains`. As a result, the
TRC is never picked up and the control service can never create a proper signer.
The only remedy would be to restart the service.

With this change, this is no longer the case.

GitOrigin-RevId: 3e994392e77946fb4df65c2af8af3cf21bdfb672