build: bazelify open API generation
Introduce a rules_openapi subfolder which contains rules for gneration
of open API bundles and Go code generation from those bundles.

The integration with gazelle is a bit itchy because gazelle doesn't have
great support for generated files, we need marker files to indicate the
package and additionally need some `#keep` comments in the build files.

Use updatesrc rules to create workspace copies of the generated files so
that editors are happy.

GitOrigin-RevId: 0bc8dfc97c226a0306745588a4b9d77889cadf04