lib/serrors: record stacktrace
Sometimes it can be quite useful to have a stacktrace in case an error
happens.

Attach the stacktrace to serrors.New and serrors.WrapStr if there isn't
already one attached. The other functions Wrap and WithCtx don't get the
treatment because they are deprecated anyway.

The stacktraces are currently only printed in the zaplogger with JSON
encoding by default. Otherwise the stacktrace would need to be manually
extracted with the StackTrace() method.

This PR is inspired by the errors library at
https://github.com/pkg/errors and also takes code from that library.

GitOrigin-RevId: a26fff0353a0117114333ada694613881413efc6