fatal: remove library
The fatal library is ill-advised. It adds another control flow outside
of regular go error handling and does not fit idiomatic go code.
It should be up to the caller to decide whether an error is "fatal"
or not. A library does not know how it is called.

For the fatal usage in the mains, we can simply use an error group to
shutdown the application if one of the important tasks errors out.

For programming mistakes, we should just use panics instead.

GitOrigin-RevId: f0866b8ae5d8978194c68e5cc2fdcc29758088f5