scion-pki: replace --healthy-only with --no-probe
Replace the `--healthy-only` flag with the `--no-probe` flag in the
`scion-pki certificate renew` command. This is more ergonomic and closer
to what a user would expect from the tool.

In the best case, we would react to SCMP messages and retry on a different
path. But probing the path first has a similar effect, and we already
have the tooling in place.

GitOrigin-RevId: cd83ca5becd868b3ad25120cbd1529dac78966cf