lib/topology: add reload functionality & drop itopo
itopo was used to reload the topology and was also designed to support
dynamic topology which we never really implemented. itopo also gave
access to the full topology which made it really hard to allow for
refactorings of the topology format. The new loader only exposes a
subset of the topology interface that is actually required by the
services.

The router integration still needs to be designed and implemented in a
later commit.

All `itopo` metrics are removed as part of this PR. There are new
`topology` metrics that offer a partial replacement:
- `topology_last_update_time`: indicates the last successful update.
- `topology_updates_total`: counts all updates and now also reports
  `err_read` in case  topology file read errors.

GitOrigin-RevId: 30c7db16e2f2a683e9ef7752f48fa990f3e4c528