router/bfd: ensure no data on received packet data
The processing of the incoming BFD messages passed a BFD message as a
`*layers.BFD` to a channel, from where it would be asynchronously be
processed. The `layers.BFD` object contains slices that reference the
underlying packet data.
While we were not accessing any of the these slice field, this was a
lurking issue. Fixed by passing a separate struct not containing any of
the slice fields through the channel.
Messages are now validated synchronously in the router dataplane, and
only the processing depending on the BFD Session state happens
asynchronously.

Moving the validation (`shouldDiscard`) to be performed synchronously in
the dataplane reveiled that this logic was not faithfully following the
specifications in the RFC 5880, Section 6.8.6 (as referenced in the doc
comment) -- the RFC states:

  > If the Your Discriminator field is zero and the State field is not
  > Down or AdminDown, the packet MUST be discarded.

In `shouldDiscard`, the `localState` was used in the corresponding test
instead of the `State` field of the received message.

Also in this change:
- remove the unused bfd.Controller. Adapt TestSessionDebootstrap to
  not use this Controller.
- reuse layers.BFD in dataplane process{intra,inter}BFD