cs: add ca disabled mode
At the moment whether CA is enabled or disabled is read from the topology file.
The CS config is a more appropriate place for the flag.

The default for the `mode` property in CS config is changed from `in-process`
to `disabled`. If you want to run in in-process mode you have to specify the
option explicitly.

GitOrigin-RevId: ce0d252a7f6654f071bec3eab5065a2a3d339af1