gateway: fix flaky TestGatewayWatcherRun test
TestGatewayWatcherRun was flaky on CI, it seems the bounds were not relaxed
enough. Instead of trying to estimate the correct bounds this commit
makes the test fully deterministic by not going into periodic execution.
While not really testing the exact same functionality, I would say this
tests enough and is worth the tradeoff because the test is no longer
flaky.

GitOrigin-RevId: f9354132b3d0fa04ef82f9b694ded5ae3d47a79b