cs: avoid head-of-line blocking in beacon propagation
Previously, beacons were sent sequentially via the beacon sender.
For many beacons + high RTT, this could result in head-of-line blocking
that would cause the beacon propagation to potentially fail.

With this change, beacons are sent in parallel over the gRPC based
beacon sender which should not have this head-of-line blocking issue
internally.

In a future change, we could think about spreading out beaconing over
time such that not all interfaces are beaconed on in the exact same
iteration.

GitOrigin-RevId: 0a66e5f246bf503394c6b6110a675eb35bbc49db