gateway: heavily refactor pathhealth monitoring
This commit changes the path health monitoring of the Gateway quite
substantially. Instead of having a single Go-routine doing path lookups
and sending probes, this commit changes the behavior to have a separate
path probing Go-routine per path and a path lookup Go-routine per
remote.

The problem with the previous approach was that the it lead to path
lookups with a timeout of 400ms which is not enough if the cache of the
local control service is expired.

With the new approach we can use a generous timeout for the path lookup
and the path probing can be done separately. Furthermore the new
approach allows to eventually have different probing intervals for
different paths.

The downside of the new approach is that we now have 2 Go-routines and 1
dispatcher connection for probes per path. This is quite extensive
however Go should be perfectly capable of dealing with a large amount of
Go-routines.

Furthermore this commit does a few improvements in the path probing that
should lead to less allocations. Namely path details are cached as well
a single snet.Packet is used for sending and receiving.

GitOrigin-RevId: 956be33074fd446a5f6e22281c73fc0d9d6b8e6b