router/bfd: reduce allocations, prometheus metrics
Use prometheus metrics directly, bypassing the metrics library.
This avoids a large amount of allocations caused by usage patterns in
the metrics library. While this does not help making progress towards a
consistent way to handle metrics, it also does not appear to make
matters worse.

Avoid various other allocations in the BFD session and sender code,
by reusing the relevant objects and avoiding temporaries to escape to
the heap.
As the BFD messages are comparably infrequent, the allocations triggered
here were usually not critical. However, for setups with a large number
of interfaces, the wasteful allocations were indeed somewhat significant
compared to the allocations from the forwarding part of the router
dataplane (which are much harder to optimize away).