addr: refactor parsing and formatting functions
This change heavily refactors the parsing and formatting functionality
in the `addr` package.

The function names are renamed to `ParseX` to be more in-line with
[netip](https://pkg.go.dev/net/netip@go1.18beta1#Addr)

Formatting and parsing formatted AS, ISD or ISD-AS numbers is moved to
functions instead of methods, as that is not an inherent property, but
rather a representational problem.

With the newly introduced FormatOptions, the behavior of formatting and
parsing can be adapted.

Use `addr.IA-parsing.patch` to upgrade the code that relies on ISD-AS parsing.

GitOrigin-RevId: d3c890f810a21e638f6280be842e6b8c414f2bd0