showpaths: error when destination is local AS
Previously, showpaths would show confusing results when the destination
is the local AS; it did list an empty path, but also showed the
error "no path alive" (because the empty path cannot be probed).
With the `-e` flag, it would also show a spurious expiration time.

Invoking showpaths with the local AS as destination cannot show anything
interesting. It seems clearest to treat this as a usage error.