slayers: recycle path when decoding
In `slayers.SCION.DecodeFromBytes`, we allocate a new path for the
correct path type on every invocation. This negates the advantages of
recycling the layer object.
Without introducing an explict "recycling bin" for paths of each path
type, we can only recycle the current path it matches the type of the
packet currently decoded. The worst case of switching back and forth
between two path types, can not be improved with this approach and this
could potentially be exploitable.