showpaths: handle empty path
Previously, showpaths would show confusing results when the destination
is the local AS; it did list an empty path, but also showed the
error "no path alive" (because the empty path cannot be probed).
With the `-e` flag, it would also show a spurious expiration time.

If the destination is the local AS, this is now specifically handled
with a message explaining that this results in an empty path.

For the json output, the `paths` field is now an empty list in this case
and so omitted.
Additionally, a `local_isd_as` (following the naming pattern from the
existing `local_ip` field) is added to the json output, to allow
consumers to detect and handle this situation.

Closes #4146

GitOrigin-RevId: 7019433b072e3b9936a762721277348ee7b53af8