refactor: HopField and InfoField as value types
`HopField` and `InfoField` have been treated as pointer types, with functions like `slayers/path/scion.Raw.GetInfoField` and `GetHopField` returning newly allocated values. These functions are used frequently in the router dataplane and as such should be considered performance sensitive and allocations should generally be avoided.

These structs are small, `InfoField` is 8 bytes, `HopField` is (now) 14 bytes. Passing these as values causes no/little stack overhead compared to passing pointers.

The main changes in this PR are:
- change `HopField.Mac` from `[]byte` to `[6]byte`, allowing the `HopField` to be considered a value type and reducing the overall memory footprint
- change `Get`/`SetInfoField`  `Get`/`SetHopField` family of functions to return/accept values instead of pointers.
- generally pass `InfoField` and `HopField` as value, except where explicitly passed as mutable reference.
- change `slayers/path/scion.Decoded` `InfoFields` from `[]*path.InfoField` to `[]path.InfoField`, same for `HopFields`. This is not strictly necessary for the other changes, but it fits the theme. This should make creating `Decoded` paths more lightweight in terms of allocations (relevant in the router only for reversing paths for SCMPs).

Some tests needed to be adapted as they used an incorrect length for the hopfield MAC, which is no longer possible with the array type for the MAC.

Closes #4143

GitOrigin-RevId: ca1c018c8d51aacf6888c1dfbeea40999b825f27