gateway: fix session monitor expiration reset
Before calling reset on a timer there should always be a stop call and
the timer's channel should be drained. Otherwise it could be that we get a
spurious timeout on the session even though there were probe replies.

This commit also makes sure that if a probe reply and the
expiration happen at the same time, we handle the probe reply
deterministically.

Finally we also make sure to close the timer and the ticker when
returning from the Run method, to release the resources.

GitOrigin-RevId: 1d138155e773a359b46be66173291b6c67eb68d1