router: check src/dst IA, check transit source IP
Add egress filtering on the source IA to mitigate accidental or
malicious source address spoofing by untrusted end-hosts.
Compring the source IA against the configured local IA for egress
packets is fast and reliable.

A malicious end host could bypass this source IA check relatively
easily; by prepending fake hop-fields, it can masquerade it's outgoing
packets as transit packets, for which the check is not applied.
To mitigate this form of spoofing attack, we add a check to for the
source IP address of transit traffic -- if this source IP doesn't match
the IP address of the corresponding sibling border router, the packet is
silently dropped.
Under reasonable the assumption that the underlying (IP-)network
infrastructure prevents untrusted end-hosts from spoofing IP addresses
of the SCION infrastructure hosts, this check prevents source IA
spoofing.

As a natural extension of the source IA checks for outgoing traffic,
we also add checks for source/destination IA for transit and incoming packets;
- for transit traffic:
  neither source nor destination IA must equal local IA
- for incoming packets:
  source IA must NOT equal local IA
  destination IA must equal local IA
- for incoming packets:
  source IA must equal local IA
  destination IA must NOT equal local IA

If any check fails, we reply with the appropriate SCMP Error
InvalidSourceAddress/InvalidDestinationAddress.
Observe that in case of InvalidSourceAddress, the correct destination IA
for the SCMP Error message is unknown. It's thus possible that the error
message will be dropped because of an invalid destination IA on the
way to the back to the originator.

We have all information needed to extend these source/destination IA
checks to packets that originate from or are destined to direct neighbors.
However, this would not provide additional tamper-resistance; if the
neighbor AS is benign, it will perform filtering on its own. In case the
neighbor AS is malicious, it could trivially bypass such a filter by
prepending fake hop fields.

Add router acceptance test cases for these new checks.
Multiple existing test cases needed to be adapted because the input
packet accidentally failed some of the new checks.