slayers: recycle path when decoding
In `slayers.SCION.DecodeFromBytes`, we allocate a new path for the
correct path type on every invocation. This negates the advantages of
recycling the layer object.

Introduce a "recycling bin" for paths that can be reused in the SCION
layer's DecodeFromBytes.
This is only used when explicitly enabled, as it only makes sense if the
layer is really reused for parsing. This opt-in also avoids issues with
tests that deep-compare the SCION layer with an expected value where
this recycling bin is empty..

Closes #4147

GitOrigin-RevId: cd7c81e4e062822bcc948eae4363c9544721f4d0