repo: move scripts/cryptoplayground to tools/
The cryptoplayground belongs to tools.

Drop scripts/mastersync.sh. This was just referred to from the
documentation, as part of the introduction to working with git.
I think the docs already explain the individual steps well enough, and
besides, there are plenty of resources elsewhere for introduction to
working with git.