Add subject flag to cert validate command
Add a `subject` flag to the `certificate validate` command.
This enables validating the IA identity of the subject of the certificate against the distinguished name.
This flag does not apply when validating a certificate chain or multiple certificates,
given that only a single certificate subject is validated.

Invalid flag usage or a conflicting subject identity leads to a validation failure.