control: only wrap errors in revhandler
Fix a bug where we would always return an error in the revhandler,
even if inserting a revocation worked without an error.

Furthermore, demote the log statement to the info level. Even if this
error happens, it is not critical to our application and does not
need to be logged at error level.

GitOrigin-RevId: 9358c3be1244ea8071dd81794cf3e7b4f95eb00b