scion-pki: add subject flag to cert verify command
Add an optional subjectIA flag to the certificate verify command.
This enables validating the IA property of the subject of the certificate against an expected value.

Invalid flag usage or a mismatching IA property of the certificate subject leads to a verification failure.

Adds a check to findIA to ensure that the ISD-AS property stored in the certificate is in the canonical string representation.
The canonical string representation of an ISD-AS is provided by the string formatting function. It is the decimal representation of the uint16 of the ISD, followed by a single dash -, followed by the string formatted representation of the AS id, which for AS ids in the BGP range is the decimal representation, and for larger AS ids consists of a colon separated hexadecimal representation of the value (see the fmtAS function in package addr).