doc: generate command documentation for scion and scion-pki tools
Add a `gendocs` command to the `scion` and `scion-pki` tools to generate a tree of RST files that describe all available commands.
Additionally, add a make target that writes the generated documentation to `doc/command`.

Add `HelpTemplates` to enable a good UI in both RTD and on the command line.

GitOrigin-RevId: 2c7356153342d81680d3316375e819b45d426da4