bazel: cleanup bazelrc, add config acceptance_all
Cleanup the .bazelrc file, in particular set bazel-remote connection for
`common` so that it also applies to the `query` subcommand, and set
sensible defaults for test outputs. Generally declutter and document
options.

Add a make target `test-acceptance` using a new bazel config
`acceptance_all` to have a simple entrypoint to run all (bazel-based)
acceptance tests, including the filetransfer demo.
For this, specify the `exclusive` tag in the acceptance test rule
definitions to enforce that tests are run sequentially -- the test
environment cannot be setup for multiple tests concurrently. Also, fix
the filetransfer demo.
Finally, fix a small mistake in tools/BUILD.bazel that caused bazel
query to report errors.