scion-pki: add --subject-isd-as flag to cert verify command
Add an optional `--subject-isd-as` flag to the certificate verify command.
This enables validating the ISD-AS property of the subject of the certificate against an expected value.
Invalid flag usage or a mismatching ISD-AS property of the certificate subject leads to a verification failure.

Add a check to findIA to ensure that the ISD-AS property stored in the certificate is in the canonical string representation.
The canonical string representation of an ISD-AS is provided by the string formatting function. It is the decimal representation of the uint16 of the ISD, followed by a single dash -, followed by the string formatted representation of the AS id, which for AS ids in the BGP range is the decimal representation, and for larger AS ids consists of a colon separated hexadecimal representation of the value (see the fmtAS function in package addr).

Closes #4179

GitOrigin-RevId: 26471b42dee55b52f5f0cb863d8deb106fd84b5d