scion-pki: use TCP/IP for cert renewal in the local AS
This commit extends the `scion-pki certs renew` command to use TCP/IP when sending the renewal request to the control service in the local AS instead of QUIC/SCION. Furthermore, cert renewal will prefer the control service on the local host if there is one present. Finally, the `--remote` flag is added that allows to specify the address of a control service to send the renewal request to. `--remote` can be specified multiple times and all remotes will be tried in order until one succeeds or all of them fail.

GitOrigin-RevId: 61a703fb30570e3b3fbea1b42d7d77c1da0ffef3