Choose the right address type when generating the SIG local topology.
There is a problem generating the local topology with the SIG configuration enabled.
When running e.g.
`scion.sh topology --sig -d topology/tiny.topo`
the address type is set to IPv4 for the SIG of the 1-ff00:0:112 AS, which is wrong.
This cases a problem by not finding a valid IP address.
Fix it by getting the right address type.