snet: return *snet.UDPAddr from snet.Conn.LocalAddr
Return the local listen address of type `*snet.UDPAddr` when invoking
`snet.Conn.LocalAddr`.

Semantically, it does not really make sense to return a *net.UDPAddr
instead of the `*snet.UDPAddr` because the ISD-AS number is also part
of the local address. Previously, there was also a asymmetry between
the local and remote address. For the remote address we would always
return an address of type `*snet.UDPAddr`.

To clearly distinguish UDP/IP and UDP/SCION addresses, the `*snet.UDPAddr.Network`
method now returns `udp/scion` instead of simply `udp` (which is already
used by `*net.UDPAddr` in the Go standard library).

This change is done in response to a very subtle bug that is observed in
applications that are part of multiple ISDs. If such an application binds
the same IP+port with different ISD-AS identities and uses both to open
QUIC connections, the second QUIC connection that is established would
fail to complete the QUIC handshake. QUIC allows to multiplex multiple
connections over the same `net.PacketConn`. For this, the QUIC library
has global state in the [multiplexer](https://github.com/lucas-clemente/quic-go/blob/0a08d62dea3397740f1e45e0dc0e50583291f2a0/multiplexer.go#L47-L56).
The map key is [calculated based on the `LocalAddr` of the underlying packet
connection](https://github.com/lucas-clemente/quic-go/blob/0a08d62dea3397740f1e45e0dc0e50583291f2a0/multiplexer.go#L68).
Because we simply returned the `*net.UDPAddr` without the ISD-AS number
in the `*snet.Conn.LocalAddr` method, both connections would map to the
same entry, even though the ISD-AS identities were different.

With the change in this commit, the problem is resolved and the
multiplexer does not have colliding entries, even in the multi-ISD
scenario.

GitOrigin-RevId: be74e81540ee9e2d3441e4b09671b3ee3a7668a6