serrors: only attach stacktrace on inner most wrapping
Only attach the stacktrace on the inner most level when wrapping.

Previously, we would attach the stacktrace to every level because we
were looking for a pointer type instead of value type.
There was also logic to check if the found `basicError` already has
a stacktrace attached. However, that would lead to every second
level getting a stacktrace attached to it. The solution is to add
the stacktrace when creating a new `basicError`, or when wrapping
an existing error that does not have yet a `basicError` in the wrapped
stack.

GitOrigin-RevId: a80d1351915cf4a176c37b9f4010fa943417cb12