build: manually fix some go_deps, use gazelle update-repos -prune
Some dependencies just won't build with gazelle without adding custom
build_directives. In this case, there are two problems:
- gazelle does not understand the `go:build ignore` instruction, causing
  it to fail to generate a BUILD.bzl file because it finds two different
  package names in one directory in qtls-go-118.
- gazelle does not really honor the release tags and so adds an
  unnecessary dependency for an internal quic-go package. This
  dependency, qtls-go-118, unsurprisingly fails to build with our
  current go-1.17, and so this unnecessary dependency breaks the build.

Adapt the Make rule for `go_deps.bzl`. Use the `-prune` flag to let
`gazelle update-repos` update the `go_deps.bzl` file, instead of
resetting and recreating it.
Moved this rule directly to the Makefile as it's now compact enough.

There was an old note (from myself) mentioning that the `-prune` flag was
ineffective. I did not observe any problems related to this.