doc: describe formal change proposal process
Describe the formal change proposal process that has been worked out for
the project governance model at the SCION Association.
This process is neither complicated nor new. It closely follows the
normal GitHub workflow with just few formalities for the decision
process of proposals sprinkled on top.

The most juicy bit, namely who is the "technical team" with the
formal decision power, is intentionally left vague for now. This
will require referencing the corresponding governance structure of the
SCION Association, which has yet to be formally established.
For now, this will just be the people with write permissions to the
repository.