dispatcher: set large send buffer size on underlay socket
Set a large send buffer size on the underlay socket in the dispatcher.
This is done to tolerate unreachable next hop routers that do not respond
to ARP requests.

Without this change, the send buffer is the same size as the unresolved
queue (packets with destinations that have not been resolved via ARP yet).
As a result, the send buffer could be filled by these packets, and legitimate
packets would get dropped.