Add Gobra to the CI and specify and verify infofield.go and hopfield.go (experimental) (#4187)
As discussed in a meeting with @shitz and @oncilla, this PR introduces experimental support for program verification in the CI of SCION, using [Gobra](https://github.com/viperproject/gobra). The idea is to evaluate how feasible it is to maintain a fraction of the SCION code-base that has been formally verified.

Initially, we focus on the files `infofield.go` and `hopfield.go` in the package `pkg/slayers/path`. These files implement fundamental data structures for the SCION implementation while being a small-enough target for our initial efforts to integrate Gobra in the CI of SCION.
Currently, we prove memory safety (incl. no accesses out-of-bounds) of the methods `SerializeTo` and `DecodeFromBytes` for both `InfoField` and `HopField`, and we prove that these methods always terminate.
There are also a couple of members marked with the annotation `//@ trusted`. These correspond to methods specified but not verified because they use some feature of Go that is still unsupported by Gobra. More concretely:
- ~~method `(*InfoField).UpdateSegID`: Gobra uses a stricter notion of Addressability at the moment and rejects the expression `hfMac[:2]`~~ (addressed)
- method `InfoField.String`: Gobra does not yet support the package `fmt`

To verify the annotated code-base, Gobra took around 3 minutes on a machine with a 2.6 GHz 6-Core Intel Core i7 and 16GB of RAM. Despite this, we expect the impact of these changes on contributions to the code base to be small given the modular nature of Gobra: Gobra verifies each method separately and our CI action caches the verification results per method.  As such, a re-verification of a specified method should only be triggered by changes to that method or to one of its dependencies.