Pass parameter by reference in `Path.Reverse` in package `onehop` (#4243)
A call to Reverse cannot modify the receiver o (not directly inside the body of Reverse or indirectly via a call to ToSCIONDecoded, given that the later copies all fields that it requires from o to a new struct). As such, it seems like a good idea to pass this by reference, reducing the amount of copying needed.

Incidentally, making this change also allows us to reduce the verification time of this method in the context of the VerifiedSCION project.