drkey: demo
Add drkey demo using the acceptance test framework.

The main component of the demo is a golang executable with the
sample code to fetch / derive DRKeys, both for a "client" and "server"
use case.

The python test runner runs this binary in a SCION topology, once for
the client and once for the server, to demonstrate (and verify) that
both sides obtain the same key.

Aside; there were some issues with unicode characters being stripped
when executing the binary using the "plumbum" library. In the output of
the demo executable, this issue caused a (very confusing) change of
units from "μs" to "s".
Fixed by switching the related invocation to use "subprocess" directly.