bazel: use hermetic python interpreter
Update the rules_python to the latest release and register the hermetic
python interpreter with version 3.10.

This should remove our reliance on the host interpreter if the python
tools are executed with bazel. E.g., when running an acceptance test,
when linting, or when running the topology generator with bazel run.