lint: bump golangci-lint version for go-1.18
Update golangci-lint version to a version that support go-1.18, in
particular the new additions to the standard libraries.
Use last version before go-1.19 support was added, i.e. one version
before golangci-lint v1.48.0. The reason is that the quic-go library
version that we currently use cannot be built with go-1.19 and so
linting fails.