lint: use docker volumes for golangci-lint cache
Use docker volumes for the module and build cache in the golangci-lint
container. Avoids having to very slowly refetch and build the all
dependencies. Ideally this would somehow reuse the caches of the bazel
golang toolchain, but I don't know if that's possible.
This adds a third, redundant cache of the entire build (one on the host,
one in bazel, and this third one for the golangci-lint cache).

This cache will typically not be effective in buildkite CI (the docker
volume goes away whenever a runner is shut down).